+++ inst/function_planning.R
@@ -7,9 +7,7 @@ religion
 family ##stats function
 sport
 
-
-
-
+for in the r_list check if it's a data frame and then return that to the list, do so the same in r_data_frame