+++ tests/testthat/test-as_integer.R
@@ -1,0 +1,7 @@
+context("Checking as_integer")
+
+test_that("as_integer ...",{
+
+
+})
+