+++ tests/testthat/test-r_dummy.R
@@ -1,0 +1,7 @@
+context("Checking r_dummy")
+
+test_that("r_dummy ...",{
+
+
+})
+