+++ tests/testthat/test-r_series.R
@@ -1,0 +1,7 @@
+context("Checking r_series")
+
+test_that("r_series ...",{
+
+
+})
+